Refactor code

+++ hibernate_springtransaction/src/main/java/springboot/hibernate/controller/InterviewController.java
@@ -6,6 +6,9 @@ import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.propertyeditors.CustomDateEditor;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.ModelMap;
 import org.springframework.web.bind.WebDataBinder;
@@ -19,7 +22,6 @@ import springboot.hibernate.service.CandidateService;
 import springboot.hibernate.service.InterviewService;
 import springboot.hibernate.service.UserService;
 
-
 @EnableWebMvc
 @Controller
 @RequestMapping(value = "/interview")
@@ -31,13 +33,13 @@ public class InterviewController {
     @Autowired
     private UserService userService;
 
-	@RequestMapping(value = "/index", method = RequestMethod.GET)
-	public String index(ModelMap modelMap) {
-		List<Interview> interviews = interviewService.findAll();
-		modelMap.addAttribute("interviewList", interviews);
+    @RequestMapping(value = "/index", method = RequestMethod.GET)
+    public String index(ModelMap modelMap) {
+        List<Interview> interviews = interviewService.findAll();
+        modelMap.addAttribute("interviewList", interviews);
 
-		return "interviewPage";
-	}
+        return "interviewPage";
+    }
 
     @RequestMapping(value = "/add/{id}", method = RequestMethod.GET)
     public String add(@PathVariable(value = "id") int id, ModelMap interviewMap, ModelMap candidateMap) {
@@ -55,7 +57,11 @@ public class InterviewController {
         interview.setCandidate(candidate);
         interview.setCandidateId(id);
 
-        User user = userService.findUserByUserName("admin");
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+
+        String name = auth.getName();
+
+        User user = userService.findUserByUserName(name);
         interview.setUser(user);
         interview.setUserId(user.getUserId());
         interviewService.create(interview);
@@ -65,32 +71,32 @@ public class InterviewController {
 
     @RequestMapping(value = "/candidateView/{id}", method = RequestMethod.GET)
     public String viewByCandidateId(@PathVariable(value = "id") int id, ModelMap modelMap) {
-	    List<Interview> interviewList = interviewService.findByCandidateId(id);
-	    modelMap.addAttribute("interviewList", interviewList);
+        List<Interview> interviewList = interviewService.findByCandidateId(id);
+        modelMap.addAttribute("interviewList", interviewList);
 
-	    return "interviewHistory";
+        return "interviewHistory";
     }
 
     @RequestMapping(value = "/userView/{id}", method = RequestMethod.GET)
     public String viewByUserId(@PathVariable(value = "id") int id, ModelMap modelMap) {
-	    List<Interview> interviewList = interviewService.findByUserId(id);
-	    modelMap.addAttribute("interviewList", interviewList);
+        List<Interview> interviewList = interviewService.findByUserId(id);
+        modelMap.addAttribute("interviewList", interviewList);
 
-	    return "interviewHistory";
+        return "interviewHistory";
     }
 
     @RequestMapping(value = "/delete/{id}", method = RequestMethod.GET)
     public String delete(@PathVariable(value = "id") int id, RedirectAttributes redirectAttributes) {
-//        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
-//        UserDetails userDetail = (UserDetails) auth.getPrincipal();
-//        Interviewer interviewer = interviewerService.findInterviewerbyUserName(userDetail.getUsername());
-//        Interview interview = interviewService.find(id);
-//        if ((interviewer.getInterviewerId() != interview.getInterviewer().getInterviewerId())
-//                && interviewer.getRolename().contains("ROLE_INTERVIEWER")) {
-//            redirectAttributes.addFlashAttribute("message",
-//                    "Your role is INTERVIEWER - You can only DELETE your candidate!!!");
-//            return "redirect:/interview.html";
-//        } else
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        UserDetails userDetail = (UserDetails) auth.getPrincipal();
+        User user = userService.findUserByUserName(userDetail.getUsername());
+        Interview interview = interviewService.find(id);
+        if ((user.getUserId() != interview.getUserId())
+                && user.getRole().getRoleName().contains("ROLE_USER")) {
+            redirectAttributes.addFlashAttribute("message",
+                    "Your role is INTERVIEWER - You can only DELETE your candidate!!!");
+            return "interviewPage";
+        } else
             interviewService.remove(id);
         redirectAttributes.addFlashAttribute("message", "The interview has been deleted successful!!!");
         return "redirect:/interview/index";
@@ -98,55 +104,28 @@ public class InterviewController {
 
     @RequestMapping(value = "/edit/{id}", method = RequestMethod.GET)
     public String edit(@PathVariable(value = "id") int id, ModelMap modelMap, RedirectAttributes redirectAttributes) {
-        // If u are interviewer, u can only edit your candidate/interview
-//        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
-//        UserDetails userDetail = (UserDetails) auth.getPrincipal();
-//        Interviewer interviewer = interviewerService.findInterviewerbyUserName(userDetail.getUsername());
+        // If u are user, u can only edit your candidate/interview
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        UserDetails userDetail = (UserDetails) auth.getPrincipal();
+        User user = userService.findUserByUserName(userDetail.getUsername());
         Interview interview = interviewService.find(id);
-//        if ((interviewer.getInterviewerId() != interview.getInterviewer().getInterviewerId())
-//                && interviewer.getRolename().contains("ROLE_INTERVIEWER")) {
-//            redirectAttributes.addFlashAttribute("message",
-//                    "Your role is INTERVIEWER - You can only EDIT your candidate!!!");
-//            return "redirect:/interview.html";
-//        } else if (interviewService.find(id) == null) {
-//            return "redirect:/403.html";
-//        } else
+        if ((user.getUserId() != interview.getUserId()) && user.getRole().getRoleName().contains("ROLE_USER")) {
+            redirectAttributes.addFlashAttribute("message", "Your role is User - You can only EDIT your candidate!!!");
+            return "redirect:/interview/index";
+        } else if (interviewService.find(id) == null) {
+            return "redirect:/403.html";
+        } else
             modelMap.addAttribute("interview", interview);
 
         return "interviewEdit";
     }
 
     @RequestMapping(value = "/edit", method = RequestMethod.POST)
-	public String editInterview(@ModelAttribute(value = "interview") Interview interview) {
-		interviewService.edit(interview);
-
+    public String editInterview(@ModelAttribute(value = "interview") Interview interview, RedirectAttributes redirectAttributes) {
+        interviewService.edit(interview);
+        redirectAttributes.addFlashAttribute("message", "The interview has been edited successful!!!");
 		return "redirect:/interview/index";
 	}
-
-//    @RequestMapping(value = "/edit", method = RequestMethod.POST)
-//    public String edit(@ModelAttribute(value = "interview") @Valid Interview interview, BindingResult bindingResult,
-//                       RedirectAttributes redirectAttributes) {
-//        InterviewValidator interviewValidator = new InterviewValidator();
-//        interviewValidator.validate(interview, bindingResult);
-//        if (bindingResult.hasErrors() || interview.getCandidate().getName() == null
-//                || interview.getCandidate().getUniName() == null || interview.getCandidate().getGpa() < 0
-//                || interview.getCandidate().getGpa() > 10 || interview.getCandidate().getSkill() == null
-//                || interview.getCandidate().getGraduateYear() == null
-//                || interview.getCandidate().getDayOfBirth() == null || interview.getCandidate().getEmail() == null
-//                || interview.getCandidate().getPhone() < 0 || interview.getCandidate().getAddress() == null
-//                || interview.getCandidate().getDegree() == null) {
-//            redirectAttributes.addFlashAttribute("message", "Please fill on the form above to edit your Interview!");
-//            return "editInterview";
-//        } else {
-//            Authentication auth = SecurityContextHolder.getContext().getAuthentication();
-//            String name = auth.getName();
-//            Interviewer i = this.interviewerService.findInterviewerbyUserName(name);
-//            interview.setInterviewer(i);
-//            interviewService.edit(interview);
-//            redirectAttributes.addFlashAttribute("message", "The interview has been edited successful!!!");
-//            return "redirect:/interview.html";
-//        }
-//    }
 
     @InitBinder
     protected void initBinder(WebDataBinder binder) {
